move posts actions from listsActions.js to a new file
